# board_study/docs/api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "board_study API",
    "description": "Spring Boot 기반 게시판 REST API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local server"
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "인증 관련 API"
    },
    {
      "name": "Post",
      "description": "게시글 관련 API"
    },
    {
      "name": "Comment",
      "description": "댓글 관련 API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SignupRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "사용자 이름 (로그인 ID)",
            "example": "user1"
          },
          "password": {
            "type": "string",
            "description": "비밀번호",
            "example": "password123"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "사용자 이름",
            "example": "user1"
          },
          "password": {
            "type": "string",
            "description": "비밀번호",
            "example": "password123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT 토큰",
            "example": "eyJ..."
          }
        }
      },
      "PostCreateRequest": {
        "type": "object",
        "required": ["title", "content"],
        "properties": {
          "title": {
            "type": "string",
            "description": "게시글 제목",
            "example": "제목입니다."
          },
          "content": {
            "type": "string",
            "description": "게시글 내용",
            "example": "내용입니다."
          }
        }
      },
      "PostUpdateRequest": {
        "type": "object",
        "required": ["title", "content"],
        "properties": {
          "title": {
            "type": "string",
            "description": "수정할 제목",
            "example": "수정된 제목입니다."
          },
          "content": {
            "type": "string",
            "description": "수정할 내용",
            "example": "수정된 내용입니다."
          }
        }
      },
      "PostListResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "게시글 ID",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "게시글 제목",
            "example": "제목입니다."
          },
          "content": {
            "type": "string",
            "description": "게시글 내용",
            "example": "내용입니다."
          },
          "username": {
            "type": "string",
            "description": "작성자",
            "example": "user1"
          },
          "createdAt": {
            "type": "string",
            "description": "작성시간 (yyyy-MM-dd HH:mm)",
            "example": "2024-01-01 12:00"
          }
        }
      },
      "PostResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "게시글 ID",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "게시글 제목",
            "example": "제목입니다."
          },
          "content": {
            "type": "string",
            "description": "게시글 내용",
            "example": "내용입니다."
          },
          "username": {
            "type": "string",
            "description": "작성자",
            "example": "user1"
          },
          "createdAt": {
            "type": "string",
            "description": "작성시간 (yyyy-MM-dd HH:mm)",
            "example": "2024-01-01 12:00"
          },
          "comments": {
            "type": "array",
            "description": "댓글 목록",
            "items": {
              "$ref": "#/components/schemas/CommentResponse"
            }
          }
        }
      },
      "PagePostListResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostListResponse"
            }
          },
          "totalPages": {
            "type": "integer",
            "description": "전체 페이지 수"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64",
            "description": "전체 게시글 수"
          },
          "number": {
            "type": "integer",
            "description": "현재 페이지 번호"
          },
          "size": {
            "type": "integer",
            "description": "페이지당 게시글 수"
          },
          "first": {
            "type": "boolean",
            "description": "첫 번째 페이지 여부"
          },
          "last": {
            "type": "boolean",
            "description": "마지막 페이지 여부"
          }
        }
      },
      "CommentCreateRequest": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "content": {
            "type": "string",
            "description": "댓글 내용",
            "example": "댓글 내용입니다."
          }
        }
      },
      "CommentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "댓글 ID",
            "example": 1
          },
          "content": {
            "type": "string",
            "description": "댓글 내용",
            "example": "댓글 내용입니다."
          },
          "username": {
            "type": "string",
            "description": "작성자",
            "example": "user1"
          },
          "createdAt": {
            "type": "string",
            "description": "작성시간 (yyyy-MM-dd HH:mm)",
            "example": "2024-01-01 12:05"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "HTTP 상태 코드",
            "example": 404
          },
          "message": {
            "type": "string",
            "description": "에러 메시지",
            "example": "게시글을 찾을 수 없습니다."
          }
        }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "회원가입",
        "description": "사용자명과 비밀번호로 회원가입한다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              },
              "example": {
                "username": "user1",
                "password": "password123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "이미 존재하는 사용자명",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "로그인",
        "description": "자격증명 검증 후 JWT 토큰을 반환한다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "username": "user1",
                "password": "password123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                },
                "example": {
                  "token": "eyJ..."
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": 401,
                  "message": "인증이 필요합니다."
                }
              }
            }
          }
        }
      }
    },
    "/post": {
      "get": {
        "tags": ["Post"],
        "summary": "게시글 목록 조회",
        "description": "게시글 목록을 페이징하여 반환한다. keyword가 있으면 제목 기반 검색을 수행한다.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "페이지 번호 (0부터 시작)",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "description": "제목 검색 키워드",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagePostListResponse"
                },
                "example": {
                  "content": [
                    {
                      "id": 1,
                      "title": "제목입니다.",
                      "content": "내용입니다.",
                      "username": "user1",
                      "createdAt": "2024-01-01 12:00"
                    }
                  ],
                  "totalPages": 1,
                  "totalElements": 1,
                  "number": 0,
                  "size": 10,
                  "first": true,
                  "last": true
                }
              }
            }
          }
        }
      }
    },
    "/post/create": {
      "post": {
        "tags": ["Post"],
        "summary": "게시글 생성",
        "description": "게시글을 생성한다.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCreateRequest"
              },
              "example": {
                "title": "제목입니다.",
                "content": "내용입니다."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "status": 401, "message": "인증이 필요합니다." }
              }
            }
          }
        }
      }
    },
    "/post/{id}": {
      "get": {
        "tags": ["Post"],
        "summary": "게시글 단건 조회",
        "description": "게시글 ID로 단건 조회한다. 댓글 목록을 포함하여 반환한다.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "게시글 ID",
            "schema": { "type": "integer", "format": "int64" },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PostResponse" },
                "example": {
                  "id": 1,
                  "title": "제목입니다.",
                  "content": "내용입니다.",
                  "username": "user1",
                  "createdAt": "2024-01-01 12:00",
                  "comments": [{ "id": 1, "content": "댓글 내용입니다.", "username": "user2", "createdAt": "2024-01-01 12:05" }]
                }
              }
            }
          },
          "404": {
            "description": "게시글 없음",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Post"],
        "summary": "게시글 수정",
        "description": "게시글을 수정한다. 작성자 본인만 가능.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "게시글 ID",
            "schema": { "type": "integer", "format": "int64" },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PostUpdateRequest" },
              "example": {
                "title": "수정된 제목입니다.",
                "content": "수정된 내용입니다."
              }
            }
          }
        },
        "responses": {
          "200": { "description": "성공" },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "status": 401, "message": "인증이 필요합니다." }
              }
            }
          },
          "403": {
            "description": "접근 권한 없음",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "status": 403, "message": "접근 권한이 없습니다." }
              }
            }
          },
          "404": {
            "description": "게시글 없음",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Post"],
        "summary": "게시글 삭제",
        "description": "게시글을 삭제한다. 작성자 본인만 가능.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "게시글 ID",
            "schema": { "type": "integer", "format": "int64" },
            "example": 1
          }
        ],
        "responses": {
          "200": { "description": "성공" },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "status": 401, "message": "인증이 필요합니다." }
              }
            }
          },
          "403": {
            "description": "접근 권한 없음",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "status": 403, "message": "접근 권한이 없습니다." }
              }
            }
          },
          "404": {
            "description": "게시글 없음",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/post/{postId}/comments": {
      "post": {
        "tags": ["Comment"],
        "summary": "댓글 작성",
        "description": "특정 게시글에 댓글을 작성한다.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "description": "게시글 ID",
            "schema": { "type": "integer", "format": "int64" },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CommentCreateRequest" },
              "example": { "content": "댓글 내용입니다." }
            }
          }
        },
        "responses": {
          "200": { "description": "성공" },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "status": 401, "message": "인증이 필요합니다." }
              }
            }
          },
          "404": {
            "description": "게시글 없음",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/post/{postId}/comments/{commentId}": {
      "delete": {
        "tags": ["Comment"],
        "summary": "댓글 삭제",
        "description": "특정 게시글의 댓글을 삭제한다. 작성자 본인만 가능.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "description": "게시글 ID",
            "schema": { "type": "integer", "format": "int64" },
            "example": 1
          },
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "description": "댓글 ID",
            "schema": { "type": "integer", "format": "int64" },
            "example": 1
          }
        ],
        "responses": {
          "200": { "description": "성공" },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "status": 401, "message": "인증이 필요합니다." }
              }
            }
          },
          "403": {
            "description": "접근 권한 없음",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "status": 403, "message": "접근 권한이 없습니다." }
              }
            }
          },
          "404": {
            "description": "게시글 또는 댓글 없음",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  }
}